add delete round, add delete match, show tournament detail

+++ src/app/api/admin/round/[tourmamentId]/[id]/delete/route.ts
@@ -4,6 +4,7 @@ import { NextRequest, NextResponse } from 'next/server'
 
 // Models: Round
 import '@/models/RoundModel'
+import MatchModel from '@/models/MatchModel'
 
 // [DELETE]: /api/admin/round/:id/delete
 export async function DELETE(req: NextRequest, { params: { id } }: { params: { id: string } }) {
@@ -12,6 +13,22 @@ export async function DELETE(req: NextRequest, { params: { id } }: { params: { i
   try {
     // connect to database
     await connectDatabase()
+
+    // get round by id
+    const round = await RoundModel.findById(id).lean()
+
+    // check if round exists
+    if (!round) {
+      return NextResponse.json({ message: 'Không tìm thấy vòng đấu' }, { status: 404 })
+    }
+
+    // get matches of round
+    const matchQuantity = await MatchModel.countDocuments({ roundId: id })
+
+    // check matches of round
+    if (matchQuantity > 0) {
+      return NextResponse.json({ message: 'Vòng đấu đang chứa trận đấu' }, { status: 400 })
+    }
 
     // delete round
     await RoundModel.findByIdAndDelete(id)
